requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.1.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
html5lib==1.1
http3==0.6.7
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.3.0
lxml==4.9.2
mechanize==0.4.8
mysql-connector-python==8.0.32
parse==1.19.0
protobuf==3.20.3
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
requests==2.28.2
requests-html==0.10.0
rfc3986==1.5.0
schedule==1.2.0
six==1.16.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
